# docs/tenant-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/outcomes": {
      "get": {
        "operationId": "OutcomesController_getOutcomes",
        "summary": "Get outcomes for current user",
        "parameters": [
          {
            "name": "weekStart",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "example": "revenue",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field and direction (e.g., createdAt:desc, name:asc)",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weekly outcomes for the tenant user"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Outcomes"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "OutcomesController_createOutcome",
        "summary": "Create new outcome",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutcomeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Outcome created"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Outcomes"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/v1/activities": {
      "get": {
        "operationId": "ActivitiesController_getAllActivities",
        "summary": "Get all activities with optional filters and pagination",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "example": "revenue",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field and direction (e.g., createdAt:desc, name:asc)",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category (Leads, Sales, Operations, etc.)",
            "schema": {
              "example": "Sales",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "example": "Active",
              "enum": [
                "Active",
                "Completed",
                "Cancelled"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filter by priority",
            "schema": {
              "example": "High",
              "enum": [
                "Low",
                "Medium",
                "High"
              ],
              "type": "string"
            }
          },
          {
            "name": "dueDateFrom",
            "required": false,
            "in": "query",
            "description": "Filter activities due on or after this date (ISO 8601)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "dueDateTo",
            "required": false,
            "in": "query",
            "description": "Filter activities due on or before this date (ISO 8601)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated activities assigned to the authenticated tenant user"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "ActivitiesController_createActivity",
        "summary": "Create activity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Activity created for current tenant"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/v1/templates/outcomes": {
      "get": {
        "operationId": "TemplatesController_getAllOutcomeTemplates",
        "summary": "Get all outcome templates",
        "description": "Returns all global outcome templates. Templates are managed by Superadmin.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ordered list of outcome templates",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "outcome_tmpl_456",
                      "title": "Publish customer newsletter",
                      "category": "Marketing",
                      "expectedImpact": "Drive engagement",
                      "createdAt": "2025-11-15T12:00:00.000Z",
                      "updatedAt": "2025-11-15T12:05:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Templates"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/templates/activities": {
      "get": {
        "operationId": "TemplatesController_getAllActivityTemplates",
        "summary": "Get all activity templates",
        "description": "Returns all global activity templates. Templates are managed by Superadmin.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ordered list of activity templates",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "activity_tmpl_789",
                      "title": "Follow up with inbound leads",
                      "category": "Sales",
                      "defaultOwnerRole": "MANAGER",
                      "createdAt": "2025-11-15T12:00:00.000Z",
                      "updatedAt": "2025-11-15T12:05:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Templates"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Accountability Tenant API",
    "description": "REST API for the BG Accountability Platform - Tenant Services.\n\n## Overview\nThis API serves both the Tenant Web Dashboard (Next.js) and Mobile App (Flutter).\n\n## Authentication\nAll protected endpoints require a valid JWT Bearer token.\n- Login via `POST /api/v1/auth/login`\n- Include token in header: `Authorization: Bearer <token>`\n- Refresh tokens via `POST /api/v1/auth/refresh`\n\n## Multi-Factor Authentication (MFA)\n- If MFA is enabled, login returns a `tempToken`\n- Complete MFA via `POST /api/v1/auth/mfa/login` with tempToken + TOTP code\n\n## Rate Limiting\n- Default: 100 requests per minute per IP\n- Auth endpoints: 10 requests per minute per IP\n\n## Error Format\nAll errors follow: `{ success: false, message: string, code?: string, details?: object }`\n\n## Pagination\nPaginated endpoints use: `{ items: T[], total: number, page: number, pageSize: number }`\n",
    "version": "1.0.0",
    "contact": {
      "name": "BG Accountability",
      "url": "https://bridgegaps.app",
      "email": "support@bridgegaps.app"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local Development"
    },
    {
      "url": "https://api.bridgegaps.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "CreateOutcomeDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Planned",
              "Done",
              "Missed"
            ]
          },
          "weekStartDate": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "weekStartDate"
        ]
      },
...
      "CreateActivityDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "default": "Active"
          }
        },
        "required": [
          "title",
          "category"
        ]
      },
...
    }
  }
}